Move some conftest up
