+++ backend/convert/chat_writer_test.py
@@ -1,12 +1,10 @@
 import pytest
-from analysis.conftest import tarsp_category
 
 from .chat_reader import ChatDocument
 from .chat_writer import ChatWriter
 
 
 @pytest.mark.django_db
-@pytest.mark.usefixtures("tarsp_category")
 def test_chat_writer(chafiles, tarsp_category):
     for inpath, outpath in chafiles:
         doc = ChatDocument.from_chatfile(inpath, tarsp_category)
